[fuzzing] use trait defaults to reduce repeated code
Closes: #5345
Approved by: bmwill
